--- src/visualizations/vis_imports.py
@@ -1,3 +1,0 @@
-from src.visualizations.visualization_interface import VisualizationInterface
-from subprocess import run
-import matplotlib.pylab as plt
